new feature in hmap gframe (=axisNB), transforming to 'hat' coordinates.
the axis given in carestian coordinates is transformed to 'hat' coordinates, which are still three fourier series but periodic on one field period. Works even for the knot.
A check is implemented and aborts if the transform of the original data is not periodic.
Makes evaluation faster compared to the full turn that was used before... especially if number of filed periods is large.