cmake flag COMPILE_SRC_PYGVEC is ON by default.
which makes pygvec fortran modules part of gveclib. when COMPILE_PYGVEC=ON, only needs  link to gveclib.
Disable pygvec flags if COMPILE_GVEC=OFF

also fix the filter for converter tests in CI, where "frenet","axisNB" and "knot" testcases are not possible to run yet (since the parameterfile is not provided).