add boundary_perturb type & depth
Added two new parameters related to specifying a boundary perturbation:
* boundary_perturb_type = 'legacy' or 'cosm'
* boundary_perturb_depth = real

type='legacy' uses the existing implementation where the boundary perturbation blends towards the axis using a gaussian. type='cosm' was added to blend using a cosine function to the power of the poloidal modenumber m. In particular this allows applying large perturbations of the m=0 mode.

depth specifies how far the perturbation should extend into the core and is per default 0.6. Previously this was fixed, but it can now also be changed for both type='legacy' and 'cosm'.